requirements.txt
Flask
flask-cors
pymysql
python-dotenv
werkzeug
requests
PyJWT==2.8.0
